most variables/functionality moved to abstract ensemble from cawpe
