+++ src/main/java/weka_uea/classifiers/ensembles/CAWPE.java
@@ -18,19 +18,13 @@ import experiments.CollateResults;
 import experiments.Experiments;
 import evaluation.MultipleClassifierEvaluation;
 import weka_uea.classifiers.ensembles.weightings.TrainAcc;
-import weka_uea.classifiers.ensembles.weightings.ModuleWeightingScheme;
 import weka_uea.classifiers.ensembles.weightings.TrainAccByClass;
 import weka_uea.classifiers.ensembles.voting.MajorityVote;
-import weka_uea.classifiers.ensembles.voting.ModuleVotingScheme;
 
 import java.io.File;
-import java.util.Arrays;
 
 import utilities.ClassifierTools;
 import evaluation.evaluators.CrossValidationEvaluator;
-import evaluation.evaluators.Evaluator;
-import evaluation.evaluators.SamplingEvaluator;
-import utilities.DebugPrinting;
 import utilities.InstanceTools;
 import weka.classifiers.Classifier;
 import weka.classifiers.bayes.BayesNet;
@@ -41,26 +35,16 @@ import weka.classifiers.meta.RotationForest;
 import weka.classifiers.trees.J48;
 import weka.classifiers.trees.RandomForest;
 import weka.core.EuclideanDistance;
-import weka.core.Instance;
 import weka.core.Instances;
 import weka.filters.SimpleBatchFilter;
-import timeseriesweka.classifiers.SaveParameterInfo;
-import evaluation.storage.ClassifierResults;
 import experiments.data.DatasetLoading;
-import java.util.concurrent.TimeUnit;
-import weka_uea.classifiers.ensembles.EnsembleModule;
 import weka_uea.classifiers.ensembles.voting.MajorityConfidence;
 import timeseriesweka.filters.SAX;
-import utilities.ErrorReport;
-import static utilities.GenericTools.indexOfMax;
-import weka.classifiers.AbstractClassifier;
 import weka.classifiers.functions.Logistic;
 import weka.classifiers.functions.MultilayerPerceptron;
 import weka.core.TechnicalInformation;
 import weka.core.TechnicalInformationHandler;
-import weka.filters.Filter;
 import weka_uea.classifiers.kNN;
-import timeseriesweka.classifiers.TrainAccuracyEstimator;
 
 /**
  * Can be constructed and will be ready for use from the default constructor like any other classifier.
@@ -73,7 +57,7 @@ import timeseriesweka.classifiers.TrainAccuracyEstimator;
 
 
  CLASSIFICATION SETTINGS:
- Default setup is defined by setupDefaultCAWPESettings(), i.e:
+ Default setup is defined by setupDefaultSettings(), i.e:
    Comps: SVML, MLP, NN, Logistic, C4.5
    Weight: TrainAcc(4) (train accuracies to the power 4)
    Vote: MajorityConfidence (summing probability distributions)
@@ -103,7 +87,7 @@ import timeseriesweka.classifiers.TrainAccuracyEstimator;
  *
  */
 
-public class CAWPE extends AbstractClassifier implements SaveParameterInfo, DebugPrinting, TrainAccuracyEstimator, TechnicalInformationHandler {
+public class CAWPE extends AbstractEnsemble implements TechnicalInformationHandler {
 
     @Override
     public TechnicalInformation getTechnicalInformation() {
@@ -120,130 +104,13 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
         return result;
     }
     
-    protected ModuleWeightingScheme weightingScheme = new TrainAcc(4);
-    protected ModuleVotingScheme votingScheme = new MajorityConfidence();
-    protected EnsembleModule[] modules;
-
-    protected int seed = 0;
-
-    protected SimpleBatchFilter transform;
-    protected Instances trainInsts;
-
-    //TrainAccuracyEstimate
-    protected boolean writeEnsembleTrainingFile = false;
-
-    protected boolean estimateEnsemblePerformance = true;
-    protected SamplingEvaluator trainEstimator = null;
     
-    protected ClassifierResults ensembleTrainResults = null;//data generated during buildclassifier if above = true
-    protected ClassifierResults ensembleTestResults = null;//data generated during testing
-
-    //saved after building so that it can be added to our test results, even if for some reason 
-    //we're not building/writing train results
-    protected long buildTime = -1; 
-    
-    //data info
-    protected int numTrainInsts;
-    protected int numAttributes;
-    protected int numClasses;
-    protected int testInstCounter = 0;
-    protected int numTestInsts = -1;
-    protected Instance prevTestInstance = null;
-
-    //results file handling
-    protected boolean readIndividualsResults = false;
-    protected boolean writeIndividualsResults = false;
-
-    protected boolean resultsFilesParametersInitialised;
-    
-    /**
-     * An annoying comprimise to deal with base classfiers that dont produce dists 
-     * while getting their train estimate. Off by default, shouldnt be turned on for 
-     * mass-experiments, intended for cases where user knows that dists are missing
-     * (for BOSS, in this case) but still just wants to get ensemble results anyway... 
-     */
-    protected boolean fillMissingDistsWithOneHotVectors; 
-    
-    /**
-     * if readResultsFilesDirectories.length == 1, all classifier's results read from that one path
-     * else, resultsPaths.length must equal classifiers.length, with each index aligning
-     * to the path to read the classifier's results from.
-     *
-     * e.g to read 2 classifiers from one directory, and another 2 from 2 different directories:
-     *
-     *     Index |  Paths  | Classifier
-     *     --------------------------
-     *       0   |  pathA  |   c1
-     *       1   |  pathA  |   c2
-     *       2   |  pathB  |   c3
-     *       3   |  pathC  |   c4
-     *
-     */
-    protected String readResultsFilesDirectories[] = null;
-
-    /**
-     * if resultsWritePath is not set, will default to resultsPaths[0]
-     * i.e, if only reading from one directory, will write back the chosen results
-     * under the same directory. if reading from multiple directories but a particular
-     * write path not set, will simply pick the first one given.
-     */
-    protected String writeResultsFilesDirectory = null;
-    protected String ensembleName = "CAWPE";
-    protected int resampleIdentifier;
-    protected String datasetName;
 
     public CAWPE() {
-        this.ensembleName = "CAWPE";
-        this.transform = null;
-        this.setupDefaultCAWPESettings();
+        super();
     }
 
-    public Classifier[] getClassifiers(){
-        Classifier[] classifiers = new Classifier[modules.length];
-        for (int i = 0; i < modules.length; i++)
-            classifiers[i] = modules[i].getClassifier();
-        return classifiers;
-    }
 
-    public void setClassifiersNamesForFileRead(String[] classifierNames) {
-        setClassifiers(null,classifierNames,null);
-
-    }
-
-    /**
-     * If building CAWPE from scratch, the minimum requirement for running is the
- classifiers array, the others could be left null.
-     *
-     * If building CAWPE from the results files of individuals, the minimum requirement for
- running is the classifierNames list.
-     *
-     * @param classifiers array of classifiers to use
-     * @param classifierNames if null, will use the classifiers' class names by default
-     * @param classifierParameters  if null, parameters of each classifier empty by default
-     */
-    public void setClassifiers(Classifier[] classifiers, String[] classifierNames, String[] classifierParameters) {
-        if (classifiers == null) {
-            classifiers = new Classifier[classifierNames.length];
-            for (int i = 0; i < classifiers.length; i++)
-                classifiers[i] = null;
-        }
-
-        if (classifierNames == null) {
-            classifierNames = new String[classifiers.length];
-            for (int i = 0; i < classifiers.length; i++)
-                classifierNames[i] = classifiers[i].getClass().getSimpleName();
-        }
-
-        if (classifierParameters == null) {
-            classifierParameters = new String[classifiers.length];
-            for (int i = 0; i < classifiers.length; i++)
-                classifierParameters[i] = "";
-        }
-
-        this.modules = new EnsembleModule[classifiers.length];
-        for (int m = 0; m < modules.length; m++)
-            modules[m] = new EnsembleModule(classifierNames[m], classifiers[m], classifierParameters[m]);
-    }
 
     /**
      * Comps: NN, SVML, SVMQ, C4.5, NB, BN, RotF, RandF
@@ -320,9 +187,12 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
      * Weight: TrainAcc(4) (train accuracies to the power 4)
      * Vote: MajorityConfidence (summing probability distributions)
      */
-    public final void setupDefaultCAWPESettings(){
+    public final void setupDefaultSettings(){
+        this.ensembleName = "CAWPE";
+        
         this.weightingScheme = new TrainAcc(4);
         this.votingScheme = new MajorityConfidence();
+        this.transform = null;
         
         CrossValidationEvaluator cv = new CrossValidationEvaluator(seed, false, false, false, false); 
         cv.setNumFolds(10);
@@ -366,7 +236,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
      * Weight: TrainAcc(4) (train accuracies to the power 4)
      * Vote: MajorityConfidence (summing probability distributions)
      */
-    public final void setupDefaultCAWPESettings_NoLogistic(){
+    public final void setupDefaultSettings_NoLogistic(){
         this.weightingScheme = new TrainAcc(4);
         this.votingScheme = new MajorityConfidence();
         
@@ -404,7 +274,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
     }
 
 
-    public final void setupAdvancedCAWPESettings(){
+    public final void setupAdvancedSettings(){
         this.weightingScheme = new TrainAcc(4);
         this.votingScheme = new MajorityConfidence();
         
@@ -438,659 +308,8 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
 
 
 
-    public void setEstimateEnsemblePerformance(boolean b) {
-        estimateEnsemblePerformance = b;
-    }
-
-    public void setRandSeed(int seed){
-        this.seed = seed;
-    }
-
-    @Override
-    public void buildClassifier(Instances data) throws Exception {        
-        printlnDebug("**CAWPE TRAIN**");
-        
-        long startTime = System.nanoTime();
-
-        //housekeeping
-        if (resultsFilesParametersInitialised) {
-            if (readResultsFilesDirectories.length > 1)
-                if (readResultsFilesDirectories.length != modules.length)
-                    throw new Exception("CAWPE.buildClassifier: more than one results path given, but number given does not align with the number of classifiers/modules.");
-
-            if (writeResultsFilesDirectory == null)
-                writeResultsFilesDirectory = readResultsFilesDirectories[0];
-        }
-        
-        //transform data if specified
-        if(this.transform==null){
-            this.trainInsts = data;
-//            this.trainInsts = new Instances(data);
-        }else{
-           transform.setInputFormat(data);
-           this.trainInsts = Filter.useFilter(data,transform);
-        }
-          
-        //init
-        this.numTrainInsts = trainInsts.numInstances();
-        this.numClasses = trainInsts.numClasses();
-        this.numAttributes = trainInsts.numAttributes();
-
-        //set up modules
-        initialiseModules();
-        
-        //if modules' results are being read in from file, ignore the i/o overhead 
-        //of loading the results, we'll sum the actual buildtimes of each module as 
-        //reported in the files
-        if (readIndividualsResults)
-            startTime = System.nanoTime();
-        
-        //set up ensemble
-        weightingScheme.defineWeightings(modules, numClasses);
-        votingScheme.trainVotingScheme(modules, numClasses);
-
-        buildTime = System.nanoTime() - startTime;
-        if (readIndividualsResults) {
-            //we need to sum the modules' reported build time as well as the weight
-            //and voting definition time
-            for (EnsembleModule module : modules) {
-                buildTime += module.trainResults.getBuildTimeInNanos();
-                buildTime += module.trainResults.getErrorEstimateTime();
-            }
-        }
-        
-        ensembleTrainResults = new ClassifierResults();
-        ensembleTrainResults.setTimeUnit(TimeUnit.NANOSECONDS);
-        
-        if(estimateEnsemblePerformance)
-            ensembleTrainResults = estimateEnsemblePerformance(data); //combine modules to find overall ensemble trainpreds
-        
-        //HACK FOR CAWPE_EXTENSION PAPER: 
-        //since experiments expects us to make a train results object 
-        //and for us to record our build time, going to record it here instead of 
-        //editting experiments to record the buildtime at that level
-        
-        //buildTime does not include the ensemble's trainEstimator in any case, only the work required to be ready for testing
-        //time unit has been set in estimateEnsemblePerformance(data);
-        ensembleTrainResults.turnOffZeroTimingsErrors();
-        ensembleTrainResults.setBuildTime(buildTime);
-        ensembleTrainResults.turnOnZeroTimingsErrors();
-        
-        if (writeEnsembleTrainingFile)
-            writeEnsembleTrainAccuracyEstimateResultsFile();
-        
-        this.testInstCounter = 0; //prep for start of testing
-    }
-
-    protected void initialiseModules() throws Exception {
-        //currently will only have file reading ON or OFF (not load some files, train the rest)
-        //having that creates many, many, many annoying issues, especially when classifying test cases
-        if (readIndividualsResults) {
-            if (!resultsFilesParametersInitialised)
-                throw new Exception("Trying to load CAWPE modules from file, but parameters for results file reading have not been initialised");
-            loadModules(); //will throw exception if a module cannot be loaded (rather than e.g training that individual instead)
-        }
-        else
-            trainModules();
-
-        
-        for (int m = 0; m < modules.length; m++) { 
-            //in case train results didnt have probability distributions, hack for old hive cote results tony todo clean
-            modules[m].trainResults.setNumClasses(trainInsts.numClasses());
-            
-            if (fillMissingDistsWithOneHotVectors)
-                modules[m].trainResults.populateMissingDists();
-                        
-            modules[m].trainResults.findAllStatsOnce();
-        }
-    }
-
-    protected void trainModules() throws Exception {
-
-        for (EnsembleModule module : modules) {
-            if (module.getClassifier() instanceof TrainAccuracyEstimator) {
-                module.getClassifier().buildClassifier(trainInsts);
-
-                //these train results should also include the buildtime
-                module.trainResults = ((TrainAccuracyEstimator)module.getClassifier()).getTrainResults();
-                module.trainResults.finaliseResults();
-                
-                if (writeIndividualsResults) { //if we're doing trainFold# file writing
-                    String params = module.getParameters();
-                    if (module.getClassifier() instanceof SaveParameterInfo)
-                        params = ((SaveParameterInfo)module.getClassifier()).getParameters();
-                    writeResultsFile(module.getModuleName(), params, module.trainResults, "train"); //write results out
-                    printlnDebug(module.getModuleName() + " writing train file data gotten through TrainAccuracyEstimate...");
-                }
-            }
-            else {
-                printlnDebug(module.getModuleName() + " estimateing performance...");
-                module.trainResults = trainEstimator.evaluate(module.getClassifier(), trainInsts);
-                module.trainResults.finaliseResults();
-                
-                //assumption: classifiers that maintain a classifierResults object, which may be the same object that module.trainResults refers to,
-                //and which this subsequent building of the final classifier would tamper with, would have been handled as an instanceof TrainAccuracyEstimate above
-                long startTime = System.nanoTime();
-                module.getClassifier().buildClassifier(trainInsts);
-                module.trainResults.setBuildTime(System.nanoTime() - startTime);
-                module.trainResults.setTimeUnit(TimeUnit.NANOSECONDS);
-
-                if (writeIndividualsResults) { //if we're doing trainFold# file writing
-                    writeResultsFile(module.getModuleName(), module.getParameters(), module.trainResults, "train"); //write results out
-                    printlnDebug(module.getModuleName() + " writing train file with full preds from scratch...");
-                }
-            }
-        }
-    }
-
-    protected void loadModules() throws Exception {
-        //will look for all files and report all that are missing, instead of bailing on the first file not found
-        //just helps debugging/running experiments a little
-        ErrorReport errors = new ErrorReport("Errors while loading modules from file. Directories given: " + Arrays.toString(readResultsFilesDirectories));
-
-        //for each module
-        for(int m = 0; m < this.modules.length; m++){
-            String readResultsFilesDirectory = readResultsFilesDirectories.length == 1 ? readResultsFilesDirectories[0] : readResultsFilesDirectories[m];
-
-            boolean trainResultsLoaded = false;
-            boolean testResultsLoaded = false;
-
-            //try and load in the train/test results for this module
-            File moduleTrainResultsFile = findResultsFile(readResultsFilesDirectory, modules[m].getModuleName(), "train");
-            if (moduleTrainResultsFile != null) {
-                printlnDebug(modules[m].getModuleName() + " train loading... " + moduleTrainResultsFile.getAbsolutePath());
-
-                modules[m].trainResults = new ClassifierResults(moduleTrainResultsFile.getAbsolutePath());
-                trainResultsLoaded = true;
-            }
-
-            File moduleTestResultsFile = findResultsFile(readResultsFilesDirectory, modules[m].getModuleName(), "test");
-            if (moduleTestResultsFile != null) {
-                //of course these results not actually used at all during training,
-                //only loaded for future use when classifying with ensemble
-                printlnDebug(modules[m].getModuleName() + " test loading..." + moduleTestResultsFile.getAbsolutePath());
-
-                modules[m].testResults = new ClassifierResults(moduleTestResultsFile.getAbsolutePath());
-
-                numTestInsts = modules[m].testResults.numInstances();
-                testResultsLoaded = true;
-            }
-
-            if (!trainResultsLoaded)
-                errors.log("\nTRAIN results files for '" + modules[m].getModuleName() + "' on '" + datasetName + "' fold '" + resampleIdentifier + "' not found. ");
-            else if (needIndividualTrainPreds() && modules[m].trainResults.getProbabilityDistributions().isEmpty())
-                errors.log("\nNo pred/distribution for instance data found in TRAIN results file for '" + modules[m].getModuleName() + "' on '" + datasetName + "' fold '" + resampleIdentifier + "'. ");
-
-            if (!testResultsLoaded)
-                errors.log("\nTEST results files for '" + modules[m].getModuleName() + "' on '" + datasetName + "' fold '" + resampleIdentifier + "' not found. ");
-            else if (modules[m].testResults.numInstances()==0)
-                errors.log("\nNo prediction data found in TEST results file for '" + modules[m].getModuleName() + "' on '" + datasetName + "' fold '" + resampleIdentifier + "'. ");
-        }
-
-        errors.throwIfErrors();
-    }
-
-    protected boolean needIndividualTrainPreds() {
-        return estimateEnsemblePerformance || weightingScheme.needTrainPreds || votingScheme.needTrainPreds;
-    }
-
-    protected File findResultsFile(String readResultsFilesDirectory, String classifierName, String trainOrTest) {
-        File file = new File(readResultsFilesDirectory+classifierName+"/Predictions/"+datasetName+"/"+trainOrTest+"Fold"+resampleIdentifier+".csv");
-        if(!file.exists() || file.length() == 0)
-            return null;
-        else
-            return file;
-    }
-
-    //hack for handling train accuracy estimate. experiments is giving us the full path and filename
-    //to write to, instead of just the folder and expecting us to fill in the +classifierName+"/Predictions/"+datasetName+filename;
-    //when doing the interface overhaul, sort this stuff out.
-    protected void writeEnsembleTrainAccuracyEstimateResultsFile() throws Exception {
-        ensembleTrainResults.writeFullResultsToFile(writeResultsFilesDirectory);
-    }
     
-    protected void writeResultsFile(String classifierName, String parameters, ClassifierResults results, String trainOrTest) throws Exception {
-        String fullPath = writeResultsFilesDirectory+classifierName+"/Predictions/"+datasetName;
-        new File(fullPath).mkdirs();
-        fullPath += "/" + trainOrTest + "Fold" + seed + ".csv";
-        
-        results.setClassifierName(classifierName);
-        results.setDatasetName(datasetName);
-        results.setFoldID(seed);
-        results.setSplit(trainOrTest);
-        
-        results.setParas(parameters);
-        results.writeFullResultsToFile(fullPath);
-    }
-
-    /**
-     * must be called (this or the directory ARRAY overload) in order to build ensemble from results files or to write individual's
-     * results files
-     *
-     * exitOnFilesNotFound defines whether the ensemble will simply throw exception/exit if results files
-     * arnt found, or will try to carry on (e.g train the classifiers normally)
-     */
-    public void setResultsFileLocationParameters(String individualResultsFilesDirectory, String datasetName, int resampleIdentifier) {
-        resultsFilesParametersInitialised = true;
-
-        this.readResultsFilesDirectories = new String [] {individualResultsFilesDirectory};
-        this.datasetName = datasetName;
-        this.resampleIdentifier = resampleIdentifier;
-    }
-
-    /**
-     * must be called (this or the single directory string overload) in order to build ensemble from results files or to write individual's
-     * results files
-     *
-     * exitOnFilesNotFound defines whether the ensemble will simply throw exception/exit if results files
-     * arnt found, or will try to carry on (e.g train the classifiers normally)
-     */
-    public void setResultsFileLocationParameters(String[] individualResultsFilesDirectories, String datasetName, int resampleIdentifier) {
-        resultsFilesParametersInitialised = true;
-
-        this.readResultsFilesDirectories = individualResultsFilesDirectories;
-        this.datasetName = datasetName;
-        this.resampleIdentifier = resampleIdentifier;
-    }
-
-    /**
-     * if writing results of individuals/ensemble, but want to define a specific folder to write to as opposed to defaulting to the (only or first)
-     * reading location
-     */
-    public void setResultsFileWritingLocation(String writeResultsFilesDirectory) {
-        this.writeResultsFilesDirectory = writeResultsFilesDirectory;
-    }
-
-    public void setBuildIndividualsFromResultsFiles(boolean b) {
-        readIndividualsResults = b;
-        if (b)
-            writeIndividualsResults = false;
-    }
-
-    public void setWriteIndividualsTrainResultsFiles(boolean b) {
-        writeIndividualsResults = b;
-        if (b)
-            readIndividualsResults = false;
-    }
     
-    protected ClassifierResults estimateEnsemblePerformance(Instances data) throws Exception {
-        double actual, pred;
-        double[] dist;
-
-        ClassifierResults trainResults = new ClassifierResults(data.numClasses());
-        trainResults.setTimeUnit(TimeUnit.NANOSECONDS);
-        
-        long estimateTimeStart = System.nanoTime();
-        
-        //for each train inst
-        for (int i = 0; i < numTrainInsts; i++) {
-            long startTime = System.nanoTime();
-            dist = votingScheme.distributionForTrainInstance(modules, i);
-            long predTime = System.nanoTime()- startTime; //time for ensemble to form vote
-            for (EnsembleModule module : modules) //                 +time for each member's predictions
-                predTime += module.trainResults.getPredictionTime(i);
-
-            pred = utilities.GenericTools.indexOfMax(dist);
-            actual = data.instance(i).classValue();
-
-            trainResults.turnOffZeroTimingsErrors();
-            trainResults.addPrediction(actual, dist, pred, predTime, "");
-            trainResults.turnOnZeroTimingsErrors();
-        }
-        
-        long estimateTime = System.nanoTime() - estimateTimeStart;
-        for (EnsembleModule module : modules)
-            estimateTime += module.trainResults.getErrorEstimateTime();
-
-        trainResults.setClassifierName(ensembleName);
-        if (datasetName == null || datasetName.equals(""))
-            datasetName = data.relationName();
-        trainResults.setDatasetName(datasetName);
-        trainResults.setFoldID(seed);
-        trainResults.setSplit("train");
-        trainResults.setParas(getParameters());
-        
-        trainResults.setErrorEstimateTime(estimateTime);
-        trainResults.setErrorEstimateMethod(modules[0].trainResults.getErrorEstimateMethod());
-        
-        trainResults.finaliseResults();
-        
-        return trainResults;
-    }
-
-    /**
-     * If building individuals from scratch, i.e not read results from files, call this
-     * after testing is complete to build each module's testResults (accessible by module.testResults)
-     *
-     * This will be done internally anyway if writeIndividualTestFiles(...) is called, this method
-     * is made public only so that results can be accessed from memory during the same run if wanted
-     */
-    public void finaliseIndividualModuleTestResults(double[] testSetClassVals) throws Exception {
-        for (EnsembleModule module : modules)
-            module.testResults.finaliseResults(testSetClassVals); //converts arraylists to double[]s and preps for writing
-    }
-
-    /**
-     * If building individuals from scratch, i.e not read results from files, call this
-     * after testing is complete to build each module's testResults (accessible by module.testResults)
-     *
-     * This will be done internally anyway if writeIndividualTestFiles(...) is called, this method
-     * is made public only so that results can be accessed from memory during the same run if wanted
-     */
-    public void finaliseEnsembleTestResults(double[] testSetClassVals) throws Exception {
-        this.ensembleTestResults.finaliseResults(testSetClassVals);
-    }
-
-    /**
-     * @param throwExceptionOnFileParamsNotSetProperly added to make experimental code smoother,
-     *  i.e if false, can leave the call to writeIndividualTestFiles(...) in even if building from file, and this
-     *  function will just do nothing. else if actually intending to write test results files, pass true
-     *  for exceptions to be thrown in case of genuine missing parameter settings
-     * @throws Exception
-     */
-    public void writeIndividualTestFiles(double[] testSetClassVals, boolean throwExceptionOnFileParamsNotSetProperly) throws Exception {
-        if (!writeIndividualsResults || !resultsFilesParametersInitialised) {
-            if (throwExceptionOnFileParamsNotSetProperly)
-                throw new Exception("to call writeIndividualTestFiles(), must have called setResultsFileLocationParameters(...) and setWriteIndividualsResultsFiles()");
-            else
-                return; //do nothing
-        }
-
-        finaliseIndividualModuleTestResults(testSetClassVals);
-
-        for (EnsembleModule module : modules)
-            writeResultsFile(module.getModuleName(), module.getParameters(), module.testResults, "test");
-    }
-
-    /**
-     * @param throwExceptionOnFileParamsNotSetProperly added to make experimental code smoother,
-     *  i.e if false, can leave the call to writeIndividualTestFiles(...) in even if building from file, and this
-     *  function will just do nothing. else if actually intending to write test results files, pass true
-     *  for exceptions to be thrown in case of genuine missing parameter settings
-     * @throws Exception
-     */
-    public void writeEnsembleTrainTestFiles(double[] testSetClassVals, boolean throwExceptionOnFileParamsNotSetProperly) throws Exception {
-        if (!resultsFilesParametersInitialised) {
-            if (throwExceptionOnFileParamsNotSetProperly)
-                throw new Exception("to call writeEnsembleTrainTestFiles(), must have called setResultsFileLocationParameters(...)");
-            else
-                return; //do nothing
-        }
-
-        if (ensembleTrainResults != null) //performed trainEstimator
-            writeResultsFile(ensembleName, getParameters(), ensembleTrainResults, "train");
-
-        this.ensembleTestResults.finaliseResults(testSetClassVals);
-        writeResultsFile(ensembleName, getParameters(), ensembleTestResults, "test");
-    }
-
-    public EnsembleModule[] getModules() {
-        return modules;
-    }
-
-    public Evaluator getTrainEstimator() {
-        return trainEstimator;
-    }
-    
-    public void setTrainEstimator(SamplingEvaluator trainEstimator) {
-        this.trainEstimator = trainEstimator;
-    }
-
-    public String[] getClassifierNames() {
-        String[] classifierNames = new String[modules.length];
-        for (int m = 0; m < modules.length; m++)
-            classifierNames[m] = modules[m].getModuleName();
-        return classifierNames;
-    }
-
-    @Override
-    public double[] getTrainPreds() {
-        return ensembleTrainResults.getPredClassValsAsArray();
-    }
-
-    @Override
-    public double getTrainAcc() {
-        return ensembleTrainResults.getAcc();
-    }
-
-    public String getEnsembleName() {
-        return ensembleName;
-    }
-
-    public void setEnsembleName(String ensembleName) {
-        this.ensembleName = ensembleName;
-    }
-
-    public boolean getFillMissingDistsWithOneHotVectors() {
-        return fillMissingDistsWithOneHotVectors;
-    }
-
-    public void setFillMissingDistsWithOneHotVectors(boolean fillMissingDistsWithOneHotVectors) {
-        this.fillMissingDistsWithOneHotVectors = fillMissingDistsWithOneHotVectors;
-    }
-    
-    public double[][] getPosteriorIndividualWeights() {
-        double[][] weights = new double[modules.length][];
-        for (int m = 0; m < modules.length; ++m)
-            weights[m] = modules[m].posteriorWeights;
-
-        return weights;
-    }
-
-    public ModuleVotingScheme getVotingScheme() {
-        return votingScheme;
-    }
-
-    public void setVotingScheme(ModuleVotingScheme votingScheme) {
-        this.votingScheme = votingScheme;
-    }
-
-    public ModuleWeightingScheme getWeightingScheme() {
-        return weightingScheme;
-    }
-
-    public void setWeightingScheme(ModuleWeightingScheme weightingScheme) {
-        this.weightingScheme = weightingScheme;
-    }
-
-    public double[] getIndividualAccEstimates() {
-        double [] accs = new double[modules.length];
-        for (int i = 0; i < modules.length; i++)
-            accs[i] = modules[i].trainResults.getAcc();
-
-        return accs;
-    }
-
-    public double[] getPriorIndividualWeights() {
-        double[] priors = new double[modules.length];
-        for (int i = 0; i < modules.length; i++)
-            priors[i] = modules[i].priorWeight;
-        
-        return priors;
-    }
-
-    public void setPriorIndividualWeights(double[] priorWeights) throws Exception {
-        if (priorWeights.length != modules.length) 
-            throw new Exception("Number of prior weights being set (" + priorWeights.length 
-                    + ") not equal to the number of modules (" + modules.length + ")");
-        
-        for (int i = 0; i < modules.length; i++)
-            modules[i].priorWeight = priorWeights[i];
-    }
-
-    private void setDefaultPriorWeights() {
-        for (int i = 0; i < modules.length; i++)
-           modules[i].priorWeight = 1.0;
-    }
-
-    public double[][] getIndividualEstimatePredictions() {
-        double [][] preds = new double[modules.length][];
-        for (int i = 0; i < modules.length; i++)
-            preds[i] = modules[i].trainResults.getPredClassValsAsArray();
-        return preds;
-    }
-
-    public SimpleBatchFilter getTransform(){
-        return this.transform;
-    }
-
-    public void setTransform(SimpleBatchFilter transform){
-        this.transform = transform;
-    }
-
-    @Override //TrainAccuracyEstimate
-    public void writeTrainEstimatesToFile(String path) {
-        estimateEnsemblePerformance=true;
-        writeEnsembleTrainingFile=true;
-        
-        setResultsFileWritingLocation(path);
-    }
-    @Override
-    public void setFindTrainAccuracyEstimate(boolean estimatePerformance){
-        estimateEnsemblePerformance=estimatePerformance;
-    }
-
-    @Override
-    public boolean findsTrainAccuracyEstimate(){ return estimateEnsemblePerformance;}
-
-    @Override
-    public ClassifierResults getTrainResults(){
-        return ensembleTrainResults;
-    }
-
-    public ClassifierResults getTestResults(){
-        return ensembleTestResults;
-    }
-
-    @Override
-    public String getParameters(){
-        StringBuilder out = new StringBuilder();
-        out.append(weightingScheme.toString()).append(",").append(votingScheme.toString()).append(",");
-
-        for(int m = 0; m < modules.length; m++){
-            out.append(modules[m].getModuleName()).append("(").append(modules[m].priorWeight);
-            for (int j = 0; j < modules[m].posteriorWeights.length; ++j)
-                out.append("/").append(modules[m].posteriorWeights[j]);
-            out.append("),");
-        }
-
-        return out.toString();
-    }
-
-//    public void readParameters(String paramLine) {
-//        String[] classifiers = paramLine.split(",");
-//
-//        String[] classifierNames = new String[classifiers.length];
-//        double[] priorWeights = new double[classifiers.length];
-//        double[] postWeights = new double[classifiers.length];
-//
-//        for (int i = 0; i < classifiers.length; ++i) {
-//            String[] parts = classifiers[i].split("(");
-//            classifierNames[i] = parts[0];
-//            String[] weights = parts[1].split("/");
-//            priorWeights[i] = Integer.parseInt(weights[0]);
-//            for (int j = 1; j < weights.length; ++j)
-//                postWeights[j-1] = Integer.parseInt(weights[j]);
-//        }
-//
-//    }
-
-    @Override
-    public double[] distributionForInstance(Instance instance) throws Exception{
-        Instance ins = instance;
-        if(this.transform!=null){
-            Instances rawContainer = new Instances(instance.dataset(),0);
-            rawContainer.add(instance);
-            transform.setInputFormat(rawContainer);
-            Instances converted = Filter.useFilter(rawContainer,transform);
-            ins = converted.instance(0);
-        }
-
-        if (ensembleTestResults == null || (testInstCounter == 0 && prevTestInstance == null)) {//definitely the first call, not e.g the first inst being classified for the second time
-            printlnDebug("\n**TEST**");
-
-            ensembleTestResults = new ClassifierResults(numClasses);
-            ensembleTestResults.setTimeUnit(TimeUnit.NANOSECONDS);
-            ensembleTestResults.setBuildTime(buildTime);
-        }
-
-        if (readIndividualsResults && testInstCounter >= numTestInsts) //if no test files loaded, numTestInsts == -1
-            throw new Exception("Received more test instances than expected, when loading test results files, found " + numTestInsts + " test cases");
-
-        double[] dist;
-        long startTime = System.nanoTime();
-        long predTime;
-        if (readIndividualsResults) { //have results loaded from file
-            dist = votingScheme.distributionForTestInstance(modules, testInstCounter);
-            predTime = System.nanoTime() - startTime; //time for ensemble to form vote
-            for (EnsembleModule module : modules) //            +time for each member's predictions
-                predTime += module.testResults.getPredictionTime(testInstCounter);
-        }
-        else {//need to classify them normally
-            dist = votingScheme.distributionForInstance(modules, ins);
-            predTime = System.nanoTime() - startTime;
-        }
-        
-        ensembleTestResults.turnOffZeroTimingsErrors();
-        ensembleTestResults.addPrediction(dist, indexOfMax(dist), predTime, "");
-        ensembleTestResults.turnOnZeroTimingsErrors();
-        
-        if (prevTestInstance != instance)
-            ++testInstCounter;
-        prevTestInstance = instance;
-
-        return dist;
-    }
-
-    @Override
-    public double classifyInstance(Instance instance) throws Exception {
-        double[] dist = distributionForInstance(instance);
-        return utilities.GenericTools.indexOfMax(dist);
-    }
-
-    /**
-     * @return the predictions of each individual module, i.e [0] = first module's vote, [1] = second...
-     */
-    public double[] classifyInstanceByConstituents(Instance instance) throws Exception{
-        Instance ins = instance;
-        if(this.transform!=null){
-            Instances rawContainer = new Instances(instance.dataset(),0);
-            rawContainer.add(instance);
-            transform.setInputFormat(rawContainer);
-            Instances converted = Filter.useFilter(rawContainer,transform);
-            ins = converted.instance(0);
-        }
-
-        double[] predsByClassifier = new double[modules.length];
-
-        for(int i=0;i<modules.length;i++)
-            predsByClassifier[i] = modules[i].getClassifier().classifyInstance(ins);
-
-        return predsByClassifier;
-    }
-
-    /**
-     * @return the distributions of each individual module, i.e [0] = first module's dist, [1] = second...
-     */
-    public double[][] distributionForInstanceByConstituents(Instance instance) throws Exception{
-        Instance ins = instance;
-        if(this.transform!=null){
-            Instances rawContainer = new Instances(instance.dataset(),0);
-            rawContainer.add(instance);
-            transform.setInputFormat(rawContainer);
-            Instances converted = Filter.useFilter(rawContainer,transform);
-            ins = converted.instance(0);
-        }
-
-        double[][] distsByClassifier = new double[this.modules.length][];
-
-        for(int i=0;i<modules.length;i++){
-            distsByClassifier[i] = modules[i].getClassifier().distributionForInstance(ins);
-        }
-
-        return distsByClassifier;
-    }
-
    
     public static void exampleCAWPEUsage() throws Exception {
         String datasetName = "ItalyPowerDemand";
@@ -1098,7 +317,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
         Instances train = DatasetLoading.loadDataNullable("c:/tsc problems/"+datasetName+"/"+datasetName+"_TRAIN");
         Instances test = DatasetLoading.loadDataNullable("c:/tsc problems/"+datasetName+"/"+datasetName+"_TEST");
 
-        //Uses predefined default settings. This is the CAWPE classifier built on 'simple' components in the paper, equivalent to setupDefaultCAWPESettings()
+        //Uses predefined default settings. This is the CAWPE classifier built on 'simple' components in the paper, equivalent to setupDefaultSettings()
         CAWPE cawpe = new CAWPE();
 
         //Setting a transform (not used in CAWPE paper, mostly for COTE/HiveCOTE or particular applications)
@@ -1117,7 +336,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
         cawpe.setVotingScheme(new MajorityVote()); //some voting schemes require dist for inst to be defined
 
         //Using predefined default settings. This is the CAWPE classifier in the paper, equivalent to default constructor
-        cawpe.setupDefaultCAWPESettings();
+        cawpe.setupDefaultSettings();
 
         int resampleID = 0;
         cawpe.setRandSeed(resampleID);
@@ -1140,105 +359,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
         cawpe.writeEnsembleTrainTestFiles(test.attributeToDoubleArray(test.classIndex()), throwExceptionOnFileParamsNotSetProperly);
     }
 
-    public String produceEnsembleReport(boolean printPreds, boolean builtFromFile) {
-        StringBuilder sb = new StringBuilder();
-
-        sb.append("CAWPE REPORT");
-        sb.append("\nname: ").append(ensembleName);
-        sb.append("\nmodules: ").append(modules[0].getModuleName());
-        for (int i = 1; i < modules.length; i++)
-            sb.append(",").append(modules[i].getModuleName());
-        sb.append("\nweight scheme: ").append(weightingScheme);
-        sb.append("\nvote scheme: ").append(votingScheme);
-        sb.append("\ndataset: ").append(datasetName);
-        sb.append("\nfold: ").append(resampleIdentifier);
-        sb.append("\ntrain acc: ").append(ensembleTrainResults.getAcc());
-        sb.append("\ntest acc: ").append(builtFromFile ? ensembleTestResults.getAcc() : "NA");
-
-        int precision = 4;
-        int numWidth = precision+2;
-        int trainAccColWidth = 8;
-        int priorWeightColWidth = 12;
-        int postWeightColWidth = 12;
-
-        String moduleHeaderFormatString = "\n\n%20s | %"+(Math.max(trainAccColWidth, numWidth))+"s | %"+(Math.max(priorWeightColWidth, numWidth))+"s | %"+(Math.max(postWeightColWidth, this.numClasses*(numWidth+2)))+"s";
-        String moduleRowHeaderFormatString = "\n%20s | %"+trainAccColWidth+"."+precision+"f | %"+priorWeightColWidth+"."+precision+"f | %"+(Math.max(postWeightColWidth, this.numClasses*(precision+2)))+"s";
-
-        sb.append(String.format(moduleHeaderFormatString, "modules", "trainacc", "priorweights", "postweights"));
-        for (EnsembleModule module : modules) {
-            String postweights = String.format("  %."+precision+"f", module.posteriorWeights[0]);
-            for (int c = 1; c < this.numClasses; c++)
-                postweights += String.format(", %."+precision+"f", module.posteriorWeights[c]);
-
-            sb.append(String.format(moduleRowHeaderFormatString, module.getModuleName(), module.trainResults.getAcc(), module.priorWeight, postweights));
-        }
-
-
-        if (printPreds) {
-            sb.append("\n\nensemble train preds: ");
-            sb.append("\ntrain acc: ").append(ensembleTrainResults.getAcc());
-            sb.append("\n");
-            for(int i = 0; i < ensembleTrainResults.numInstances();i++)
-                sb.append(produceEnsemblePredsLine(true, i)).append("\n");
-
-            sb.append("\n\nensemble test preds: ");
-            sb.append("\ntest acc: ").append(builtFromFile ? ensembleTestResults.getAcc() : "NA");
-            sb.append("\n");
-            for(int i = 0; i < ensembleTestResults.numInstances();i++)
-                sb.append(produceEnsemblePredsLine(false, i)).append("\n");
-        }
-
-        return sb.toString();
-    }
-
-    /**
-     * trueClassVal,predClassVal,[empty],dist1,...,distC,#indpreddist1,...,indpreddistC,#module1pred,...,moduleMpred
-     * split on "#"
-     * [0] = normal results file format (true class, pred class, distforinst)
-     * [1] = number of individual unweighted votes per class
-     * [2] = the unweighted prediction of each module
-     */
-    private String produceEnsemblePredsLine(boolean train, int index) {
-        StringBuilder sb = new StringBuilder();
-
-        if (train) //pred
-            sb.append(modules[0].trainResults.getTrueClassValue(index)).append(",").append(ensembleTrainResults.getPredClassValue(index)).append(",");
-        else
-            sb.append(modules[0].testResults.getTrueClassValue(index)).append(",").append(ensembleTestResults.getPredClassValue(index)).append(",");
-
-        if (train){ //dist
-            double[] pred=ensembleTrainResults.getProbabilityDistribution(index);
-            for (int j = 0; j < pred.length; j++)
-                sb.append(",").append(pred[j]);
-        }
-        else{
-            double[] pred=ensembleTestResults.getProbabilityDistribution(index);
-            for (int j = 0; j < pred.length; j++)
-                sb.append(",").append(pred[j]);
-        }
-        sb.append(",");
-
-
-        double[] predDist = new double[numClasses]; //indpreddist
-        for (int m = 0; m < modules.length; m++) {
-            if (train)
-                ++predDist[(int)modules[m].trainResults.getPredClassValue(index)];
-            else
-                ++predDist[(int)modules[m].testResults.getPredClassValue(index)];
-        }
-        for (int c = 0; c < numClasses; c++)
-            sb.append(",").append(predDist[c]);
-        sb.append(",");
-
-        for (int m = 0; m < modules.length; m++) {
-            if (train)
-                sb.append(",").append(modules[m].trainResults.getPredClassValue(index));
-            else
-                sb.append(",").append(modules[m].testResults.getPredClassValue(index));
-        }
-
-        return sb.toString();
-    }
+    
 
     
     
@@ -1252,96 +373,6 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
     
     
     
-    
-    
-    
-    protected static void testBuildingInds(int testID) throws Exception {
-        System.out.println("testBuildingInds()");
-
-        (new File("C:/JamesLPHD/CAWPETests"+testID+"/")).mkdirs();
-
-        int numFolds = 5;
-
-        for (int fold = 0; fold < numFolds; fold++) {
-            String dataset = "breast-cancer-wisc-prog";
-    //        String dataset = "ItalyPowerDemand";
-
-            Instances all = DatasetLoading.loadDataNullable("C:/UCI Problems/"+dataset+"/"+dataset);
-    //        Instances train = ClassifierTools.loadDataThrowable("C:/tsc problems/"+dataset+"/"+dataset+"_TRAIN");
-    //        Instances test = ClassifierTools.loadDataThrowable("C:/tsc problems/"+dataset+"/"+dataset+"_TEST");
-
-            Instances[] insts = InstanceTools.resampleInstances(all, fold, 0.5);
-            Instances train = insts[0];
-            Instances test = insts[1];
-
-            CAWPE cawpe = new CAWPE();
-            cawpe.setResultsFileLocationParameters("C:/JamesLPHD/CAWPETests"+testID+"/", dataset, fold);
-            cawpe.setWriteIndividualsTrainResultsFiles(true);
-            cawpe.setEstimateEnsemblePerformance(true); //now defaults to true
-            cawpe.setRandSeed(fold);
-
-            cawpe.buildClassifier(train);
-
-            double acc = .0;
-            for (Instance instance : test) {
-                if (instance.classValue() == cawpe.classifyInstance(instance))
-                    acc++;
-            }
-            acc/=test.numInstances();
-
-            cawpe.writeIndividualTestFiles(test.attributeToDoubleArray(test.classIndex()), true);
-            cawpe.writeEnsembleTrainTestFiles(test.attributeToDoubleArray(test.classIndex()), true);
-
-            System.out.println("TrainAcc="+cawpe.getTrainResults().getAcc());
-            System.out.println("TestAcc="+cawpe.getTestResults().getAcc());
-        }
-    }
-
-    protected static void testLoadingInds(int testID) throws Exception {
-        System.out.println("testBuildingInds()");
-
-        (new File("C:/JamesLPHD/CAWPETests"+testID+"/")).mkdirs();
-
-        int numFolds = 5;
-
-        for (int fold = 0; fold < numFolds; fold++) {
-            String dataset = "breast-cancer-wisc-prog";
-    //        String dataset = "ItalyPowerDemand";
-
-            Instances all = DatasetLoading.loadDataNullable("C:/UCI Problems/"+dataset+"/"+dataset);
-    //        Instances train = ClassifierTools.loadDataThrowable("C:/tsc problems/"+dataset+"/"+dataset+"_TRAIN");
-    //        Instances test = ClassifierTools.loadDataThrowable("C:/tsc problems/"+dataset+"/"+dataset+"_TEST");
-
-            Instances[] insts = InstanceTools.resampleInstances(all, fold, 0.5);
-            Instances train = insts[0];
-            Instances test = insts[1];
-
-            CAWPE cawpe = new CAWPE();
-            cawpe.setResultsFileLocationParameters("C:/JamesLPHD/CAWPETests"+testID+"/", dataset, fold);
-            cawpe.setBuildIndividualsFromResultsFiles(true);
-            cawpe.setEstimateEnsemblePerformance(true); //now defaults to true
-            cawpe.setRandSeed(fold);
-
-            cawpe.buildClassifier(train);
-
-            double acc = .0;
-            for (Instance instance : test) {
-                if (instance.classValue() == cawpe.classifyInstance(instance))
-                    acc++;
-            }
-            acc/=test.numInstances();
-            cawpe.finaliseEnsembleTestResults(test.attributeToDoubleArray(test.classIndex()));
-
-            System.out.println("TrainAcc="+cawpe.getTrainResults().getAcc());
-            System.out.println("TestAcc="+cawpe.getTestResults().getAcc());
-        }
-    }
-
-
-
-
-
-
 
 
     /**
@@ -1410,7 +441,7 @@ public class CAWPE extends AbstractClassifier implements SaveParameterInfo, Debu
         String[] dataHeaders = { "UCI", };
         String[] dataPaths = { "C:/UCI Problems/", };
         String[][] datasets = { { "hayes-roth", "pittsburg-bridges-T-OR-D", "teaching", "wine" } };
-        String writePathBase = "C:/Temp/CAWPEReproducabiltyTests/CAWPEReproducabiltyTest007/";
+        String writePathBase = "C:/Temp/CAWPEReproducabiltyTests/CAWPEReproducabiltyTest008/";
         String writePathResults =  writePathBase + "Results/";
         String writePathAnalysis =  writePathBase + "Analysis/";
         int numFolds = 5;
